moved get_index_list
