+++ src/haddock/modules/analysis/alascan/__init__.py
@@ -2,7 +2,7 @@
 from pathlib import Path
 
 from haddock import log
-from haddock.libs.libparallel import Scheduler
+from haddock.libs.libparallel import Scheduler, get_index_list
 from haddock.libs.libutil import parse_ncores
 from haddock.modules import BaseHaddockModule
 from haddock.modules.analysis.alascan.scan import (
@@ -11,7 +11,6 @@ from haddock.modules.analysis.alascan.scan import (
     alascan_cluster_analysis,
     create_alascan_plots,
     generate_alascan_output,
-    get_index_list,
     )
